feat: updated E2E tests for main using new test docker container and types package
