+++ testdata/mock_relays.json
@@ -1,6 +1,6 @@
 [
   {
-    "applicationPublicKey": "test_app_pub_key_71849c02dc71c4859b113c8b7b66dd2fbde268381e07a7c",
+    "applicationPublicKey": "test_11b8d394ca331d7c7a71ca1896d630f6",
     "nodePublicKey": "test_node_pub_key_02fbcfbad0777942c1da5425bf0105546e7e7f53fffad9",
     "method": "eth_chainId",
     "blockchain": "49",
@@ -9,7 +9,7 @@
     "elapsedTime": 0.16475
   },
   {
-    "applicationPublicKey": "test_app_pub_key_68513268a376ed2204104b2ac4498a35d3fe26450460fb7",
+    "applicationPublicKey": "test_11b8d394ca331d7c7a71ca1896d630f6",
     "nodePublicKey": "test_node_pub_key_29f42f262efa514fbe93af1e963b40c5cddb2200dd9e0a",
     "method": "eth_getBalance",
     "blockchain": "49",
@@ -18,7 +18,7 @@
     "elapsedTime": 0.20045
   },
   {
-    "applicationPublicKey": "test_app_pub_key_756188316196fbf1018dbec087f6caee76bb4bc2861d46c",
+    "applicationPublicKey": "test_11b8d394ca331d7c7a71ca1896d630f6",
     "nodePublicKey": "test_node_pub_key_abfe693936a467d31e3246e32939ec04f5509315f39ef1",
     "method": "eth_getBlockByNumber",
     "blockchain": "49",
@@ -27,7 +27,7 @@
     "elapsedTime": 0.0813666666666667
   },
   {
-    "applicationPublicKey": "test_app_pub_key_902693c0fc1d3569207a4f6f380d3d4d514e7fd819a69d4",
+    "applicationPublicKey": "test_11b8d394ca331d7c7a71ca1896d630f6",
     "nodePublicKey": "test_node_pub_key_277ed47d71a69e1aef94f8338f859af46d57840060f17d",
     "method": "eth_getCode",
     "blockchain": "3",
@@ -36,7 +36,7 @@
     "elapsedTime": 0.15785
   },
   {
-    "applicationPublicKey": "test_app_pub_key_bba4ba1984e0c521c08874c9519cc6dcd518a15dd241f53",
+    "applicationPublicKey": "test_11b8d394ca331d7c7a71ca1896d630f6",
     "nodePublicKey": "test_node_pub_key_f64fe10e37173c0e7117490291c11e81d8851175e7ad4d",
     "method": "eth_getTransactionCount",
     "blockchain": "3",
@@ -45,7 +45,7 @@
     "elapsedTime": 0.054675
   },
   {
-    "applicationPublicKey": "test_app_pub_key_ebd3cb3076f3a5a3228c650e374c69b3665b0e5b0017a69",
+    "applicationPublicKey": "test_96c981db344ab6920b7e87853838e285",
     "nodePublicKey": "test_node_pub_key_498b2b95ba2db09a263314e5fe23ce94fbd555d23578c1",
     "method": "eth_getStorageAt",
     "blockchain": "3",
@@ -54,7 +54,7 @@
     "elapsedTime": 0.1093
   },
   {
-    "applicationPublicKey": "test_app_pub_key_bd0de0319771cfdb3b2a28d4d64e33a3a23e6dcd6057f55",
+    "applicationPublicKey": "test_96c981db344ab6920b7e87853838e285",
     "nodePublicKey": "test_node_pub_key_66b14f98b70d7e77572a6c72db611ebe28cf75d9cf542f",
     "method": "eth_blockNumber",
     "blockchain": "4",
@@ -63,7 +63,7 @@
     "elapsedTime": 0.2205
   },
   {
-    "applicationPublicKey": "test_app_pub_key_ca7fe9df6499e569ae2f4910708f636beb1efa20cebda4d",
+    "applicationPublicKey": "test_96c981db344ab6920b7e87853838e285",
     "nodePublicKey": "test_node_pub_key_b5c0fd8910bb7a121686865407e1a3ba40dcfb700d5e87",
     "method": "eth_getBalance",
     "blockchain": "4",
@@ -72,7 +72,7 @@
     "elapsedTime": 0.0932
   },
   {
-    "applicationPublicKey": "test_app_pub_key_c77cd6d8bca9d1ff21b1ad5328ba04a5a610ba1bd72e1c5",
+    "applicationPublicKey": "test_96c981db344ab6920b7e87853838e285",
     "nodePublicKey": "test_node_pub_key_124d8dcca024f05e01fb56c91800480926d62da44b0eee",
     "method": "eth_blockNumber",
     "blockchain": "4",
@@ -81,7 +81,7 @@
     "elapsedTime": 0.1162
   },
   {
-    "applicationPublicKey": "test_app_pub_key_c65ffcd508b068ea2651e5d4bd5f4760bd2643584b9ef6d",
+    "applicationPublicKey": "test_96c981db344ab6920b7e87853838e285",
     "nodePublicKey": "test_node_pub_key_dbd1105f79c1c40c2fc8d8749ea5760700cff6a0504016",
     "method": "eth_blockNumber",
     "blockchain": "49",
